hirst-painting/pip_freeze_requirements.txt
colorgram.py==1.2.0
Pillow==8.2.0
